Refactor Flask App (#14)
* refactored flask app

* adjusted README to adhere to new structure

* update dockerfile to reflect changes to refactoring

* fixed JSON import error

* fixed login response JSON serialization and removed authorization_code workflow

* cleaned up extraneous comments/imports
