+++ backend/app/models.py
@@ -1,3 +1,4 @@
+from flask import jsonify
 from flask_login import UserMixin 
 from mongoengine import Document, StringField
 
@@ -6,6 +7,24 @@ class User(UserMixin, Document):
     first_name = StringField(max_length = 30)
     last_name = StringField(max_length = 50)
     email = StringField(max_length=30)
+
+    def toJSON(self):
+        return jsonify({
+            "first_name": self.first_name,
+            "last_name": self.last_name,
+            "email": self.email
+        })
+
+class Response:
+    def __init__(self, code, message):
+        self.response_code = code
+        self.response_message = message 
+    
+    def toJSON(self):
+        return jsonify({
+            "response_code": self.response_code,
+            "response_message": self.response_message
+        })
 
 """
 class Credential(Document):
